ai-service/requirements.txt
flask
flask-limiter
groq
python-dotenv